Split codebase into appose-java and appose-python
This codebase is now appose-java. For appose-python, see:
    https://github.com/apposed/appose-python

The main immediate rationale for this change is to accommodate
JetBrains IDEs, which stomp on one another's metadata; see e.g.:
    https://stackoverflow.com/q/68829992/1207769

Having two separate codebases is admittedly more scalable and
future-proof if/when additional implementations are created. We will
no longer need any clever tricks to overlay additional languages such
as JavaScript on top of the toplevel directory of a single repository.

The main downside is that it adds complexity to project maintenance:
there are now two separate issue trackers, two separate CIs, two
separate sets of release tags, etc. But everything will be OK! Maybe.