Be defensive when encoding worker responses
The response includes data from the worker script, which means
we cannot be certain which types of objects might be present.
And some types of objects cannot be encoded natively by the JSON
encoder. Therefore, let's be careful when transmitting the response,
and if something goes wrong, issue a failure message.

See apposed/appose#8.