Make shm the third, not first, NDArray attribute
So that it matches the Python implementation of Appose.
The reason it's third in Python is so that it can have the default
parameter value of None, to provide feature parity with the Java
version's two-argument NDAray constructor that creates the SharedMemory.

And create the memory even if the three-argument constructor is used.