Increase ApposeTest#testServiceStartupFailure info
Yeah, I know it's failing because it's using the system python on the
system path. I guess ProcessBuilder does that, similar to a shell.
But I wanted to add the debug output just to be friendly and clear.
The question is: how do we configure ProcessBuilder not to do that?