Fix shm test failures on macOS
On macOS, the size of the allocated shared memory block might not
precisely match the request. But it needs to be at least as large as
the requested size. E.g.: sizes less than 16384 round up to 16384.

You might think it would be a good idea to make the size() method
of SharedMemory return the actually requested size, rather than the
weird rounded up size, but that's how it works on macOS on Python,
and we're trying to make the Java implementation match the Python
behavior as closely as possible, so... ¯\_(ツ)_/¯