Add support for setting environment variables
When building an environment, a natural thing to want to do is to
set environment variables to particular values. This change adds
methods to the environment builder API to pass environment variable
key/value pairs, which will then be passed to every worker process
later spawned by that environment. There are many reasons this can
be useful; for example, one could set configuration flags needed by
worker tasks, such as DEBUG=1 to enable generation of debug logs.

This commit also adds a Environments.envVars(String...) method for
constructing a map of current environment variable keys/values with a
convenient one-liner, in case the service process wants to pass along
a subset of its environment to any of its managed environments.