Propagate running Java's java.home to system envs
Otherwise, when launching `groovy()` or other JVM-driven workers, the
first java on the system path will be used, which may not be the same
one as the running Java. In particular, as of this writing, on systems
with an up-to-date Java installation, it might be OpenJDK 25, which does
not currently work with the Groovy 4.0.23 currently used by Appose:

    $ mvn -Dappose.debug -Dtest="ApposeTest#testScopeGroovy" test
    ...
    [INFO] Running org.apposed.appose.ApposeTest
    [SERVICE-0] {"task":"b1283524-1275-4d68-bf39-bf4212870940","requestType":"EXECUTE","inputs":{"age":100},"script":"sqrt_age = age -> {\n  return Math.sqrt(age)\n}\ntask.export(sqrt_age: sqrt_age)\nreturn sqrt_age(age)\n","queue":null}
    [SERVICE-0] {"responseType":"LAUNCH","task":"b1283524-1275-4d68-bf39-bf4212870940"}
    [WORKER-0] Exception in thread "Appose-b1283524-1275-4d68-bf39-bf4212870940" BUG! exception in phase 'semantic analysis' in source unit 'Script1.groovy' Unsupported class file major version 69
    ...
    [WORKER-0] Caused by: java.lang.IllegalArgumentException: Unsupported class file major version 69
    [WORKER-0]     at groovyjarjarasm.asm.ClassReader.<init>(ClassReader.java:200)

Of course, we could upgrade the Groovy dependency version, but it seems
more correct and less volatile across systems to just test Appose Groovy
workers with the same version of Java used to build.