Fix conflation of Builder#name methods
* Builder#name() - the *environment type* the builder produces.
* Builder#name(String) - the *desired environment name* to build.

These are two different ideas. To reduce confusion and avoid name
clashes (especially in appose-python), this commit renames the
Builder#name() method to Builder#envType(). The Builder#name(String)
method is left as is (instead of renaming to Builder#envName(String)),
with the rationale that we want fluid builder syntax to be as succinct
as possible, and it can be assumed that the thing we are naming is the
environment to be built.